refactor: catalog feature to version 2, replacing old cubit and state management with a new structure. Implement home page with recent video gallery and category sections. Remove deprecated files and update routing to reflect new architecture.
