--- lib/catalogs/catalogs_v2/cubit/catalog_state.freezed.dart
@@ -1,157 +1,0 @@
-// dart format width=80
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'catalog_state.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-// dart format off
-T _$identity<T>(T value) => value;
-/// @nodoc
-mixin _$CatalogState {
-
- CatalogStatus get status; List<VideoViewModel> get videos; bool get hasReachedMax; String? get errorMessage;
-/// Create a copy of CatalogState
-/// with the given fields replaced by the non-null parameter values.
-@JsonKey(includeFromJson: false, includeToJson: false)
-@pragma('vm:prefer-inline')
-$CatalogStateCopyWith<CatalogState> get copyWith => _$CatalogStateCopyWithImpl<CatalogState>(this as CatalogState, _$identity);
-
-
-
-@override
-bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is CatalogState&&(identical(other.status, status) || other.status == status)&&const DeepCollectionEquality().equals(other.videos, videos)&&(identical(other.hasReachedMax, hasReachedMax) || other.hasReachedMax == hasReachedMax)&&(identical(other.errorMessage, errorMessage) || other.errorMessage == errorMessage));
-}
-
-
-@override
-int get hashCode => Object.hash(runtimeType,status,const DeepCollectionEquality().hash(videos),hasReachedMax,errorMessage);
-
-@override
-String toString() {
-  return 'CatalogState(status: $status, videos: $videos, hasReachedMax: $hasReachedMax, errorMessage: $errorMessage)';
-}
-
-
-}
-
-/// @nodoc
-abstract mixin class $CatalogStateCopyWith<$Res>  {
-  factory $CatalogStateCopyWith(CatalogState value, $Res Function(CatalogState) _then) = _$CatalogStateCopyWithImpl;
-@useResult
-$Res call({
- CatalogStatus status, List<VideoViewModel> videos, bool hasReachedMax, String? errorMessage
-});
-
-
-
-
-}
-/// @nodoc
-class _$CatalogStateCopyWithImpl<$Res>
-    implements $CatalogStateCopyWith<$Res> {
-  _$CatalogStateCopyWithImpl(this._self, this._then);
-
-  final CatalogState _self;
-  final $Res Function(CatalogState) _then;
-
-/// Create a copy of CatalogState
-/// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? status = null,Object? videos = null,Object? hasReachedMax = null,Object? errorMessage = freezed,}) {
-  return _then(_self.copyWith(
-status: null == status ? _self.status : status // ignore: cast_nullable_to_non_nullable
-as CatalogStatus,videos: null == videos ? _self.videos : videos // ignore: cast_nullable_to_non_nullable
-as List<VideoViewModel>,hasReachedMax: null == hasReachedMax ? _self.hasReachedMax : hasReachedMax // ignore: cast_nullable_to_non_nullable
-as bool,errorMessage: freezed == errorMessage ? _self.errorMessage : errorMessage // ignore: cast_nullable_to_non_nullable
-as String?,
-  ));
-}
-
-}
-
-
-/// @nodoc
-
-
-class _CatalogState implements CatalogState {
-  const _CatalogState({this.status = CatalogStatus.initial, final  List<VideoViewModel> videos = const <VideoViewModel>[], this.hasReachedMax = false, this.errorMessage}): _videos = videos;
-  
-
-@override@JsonKey() final  CatalogStatus status;
- final  List<VideoViewModel> _videos;
-@override@JsonKey() List<VideoViewModel> get videos {
-  if (_videos is EqualUnmodifiableListView) return _videos;
-  // ignore: implicit_dynamic_type
-  return EqualUnmodifiableListView(_videos);
-}
-
-@override@JsonKey() final  bool hasReachedMax;
-@override final  String? errorMessage;
-
-/// Create a copy of CatalogState
-/// with the given fields replaced by the non-null parameter values.
-@override @JsonKey(includeFromJson: false, includeToJson: false)
-@pragma('vm:prefer-inline')
-_$CatalogStateCopyWith<_CatalogState> get copyWith => __$CatalogStateCopyWithImpl<_CatalogState>(this, _$identity);
-
-
-
-@override
-bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _CatalogState&&(identical(other.status, status) || other.status == status)&&const DeepCollectionEquality().equals(other._videos, _videos)&&(identical(other.hasReachedMax, hasReachedMax) || other.hasReachedMax == hasReachedMax)&&(identical(other.errorMessage, errorMessage) || other.errorMessage == errorMessage));
-}
-
-
-@override
-int get hashCode => Object.hash(runtimeType,status,const DeepCollectionEquality().hash(_videos),hasReachedMax,errorMessage);
-
-@override
-String toString() {
-  return 'CatalogState(status: $status, videos: $videos, hasReachedMax: $hasReachedMax, errorMessage: $errorMessage)';
-}
-
-
-}
-
-/// @nodoc
-abstract mixin class _$CatalogStateCopyWith<$Res> implements $CatalogStateCopyWith<$Res> {
-  factory _$CatalogStateCopyWith(_CatalogState value, $Res Function(_CatalogState) _then) = __$CatalogStateCopyWithImpl;
-@override @useResult
-$Res call({
- CatalogStatus status, List<VideoViewModel> videos, bool hasReachedMax, String? errorMessage
-});
-
-
-
-
-}
-/// @nodoc
-class __$CatalogStateCopyWithImpl<$Res>
-    implements _$CatalogStateCopyWith<$Res> {
-  __$CatalogStateCopyWithImpl(this._self, this._then);
-
-  final _CatalogState _self;
-  final $Res Function(_CatalogState) _then;
-
-/// Create a copy of CatalogState
-/// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? status = null,Object? videos = null,Object? hasReachedMax = null,Object? errorMessage = freezed,}) {
-  return _then(_CatalogState(
-status: null == status ? _self.status : status // ignore: cast_nullable_to_non_nullable
-as CatalogStatus,videos: null == videos ? _self._videos : videos // ignore: cast_nullable_to_non_nullable
-as List<VideoViewModel>,hasReachedMax: null == hasReachedMax ? _self.hasReachedMax : hasReachedMax // ignore: cast_nullable_to_non_nullable
-as bool,errorMessage: freezed == errorMessage ? _self.errorMessage : errorMessage // ignore: cast_nullable_to_non_nullable
-as String?,
-  ));
-}
-
-
-}
-
-// dart format on
