+++ lib/catalogs/home/cubit/catalog_state.dart
@@ -9,8 +9,10 @@ enum CatalogStatus { initial, loading, success, failure }
 class CatalogState with _$CatalogState {
   const factory CatalogState({
     @Default(CatalogStatus.initial) CatalogStatus status,
-    @Default(<VideoViewModel>[]) List<VideoViewModel> videos,
-    @Default(false) bool hasReachedMax,
+    @Default(<CatalogType, List<VideoViewModel>>{})
+    Map<CatalogType, List<VideoViewModel>> items,
+    @Default(<VideoViewModel>[]) List<VideoViewModel> recent,
+    @Default({}) Map<CatalogType, bool> reachedMax,
     String? errorMessage,
   }) = _CatalogState;
