+++ lib/catalogs/index.dart
@@ -1,1 +1,1 @@
-export 'catalogs/view/catalog_page.dart';
+export 'home/view/catalog_page.dart';
